chore: se implementa navegación
