+++ src/components/providers/index.ts
@@ -1,1 +1,2 @@
+export * from './HttpProvider';
 export * from './NavigationProvider';
