+++ src/navigation/MainBottomTabNavigation.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react';
-import { NavigationContainer } from '@react-navigation/native';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import Icon from '@react-native-vector-icons/material-design-icons';
 import { DashboardStackNavigation } from './DashboardStackNavigation';
 import { TransactionsStackNavigation } from './TransactionsStackNavigation';
 import { AccountsStackNavigation } from './AccountsStackNavigation';
@@ -16,51 +16,64 @@ export const MainBottomTabNavigation: React.FC = () => {
       name: "DashboardTab",
       component: DashboardStackNavigation,
       options: {
-        title: 'Inicio'
-      }
-    },
-    {
-      name: "TransactionsTab",
-      component: TransactionsStackNavigation,
-      options: {
-        title: 'Transacciones'
-      }
+        title: 'Inicio',
+        tabBarIcon: ({ color, size }) => (
+          <Icon name="home" color={color} size={size} />
+        ),
+      },
     },
     {
       name: "AccountsTab",
       component: AccountsStackNavigation,
       options: {
-        title: 'Cuentas'
-      }
+        title: 'Cuentas',
+        tabBarIcon: ({ color, size }) => (
+          <Icon name="bank" color={color} size={size} />
+        ),
+      },
+    },
+    {
+      name: "TransactionsTab",
+      component: TransactionsStackNavigation,
+      options: {
+        title: 'Transacciones',
+        tabBarIcon: ({ color, size }) => (
+          <Icon name="swap-horizontal" color={color} size={size} />
+        ),
+      },
     },
     {
       name: "ReportsTab",
       component: ReportsStackNavigation,
       options: {
-        title: 'Reportes'
-      }
+        title: 'Reportes',
+        tabBarIcon: ({ color, size }) => (
+          <Icon name="chart-bar" color={color} size={size} />
+        ),
+      },
     },
     {
       name: "SettingsTab",
       component: SettingsStackNavigation,
       options: {
-        title: 'Más'
-      }
-    }
+        title: 'Configuración',
+        tabBarIcon: ({ color, size }) => (
+          <Icon name="cog" color={color} size={size} />
+        ),
+      },
+    },
   ]), []);
 
   return (
-    <NavigationContainer>
-      <Tab.Navigator>
-        {tabRoutes.map((route) => (
-          <Tab.Screen
-            key={route.name}
-            name={route.name}
-            component={route.component}
-            options={route.options}
-          />
-        ))}
-      </Tab.Navigator>
-    </NavigationContainer>
+    <Tab.Navigator>
+      {tabRoutes.map((route) => (
+        <Tab.Screen
+          key={route.name}
+          name={route.name}
+          component={route.component}
+          options={route.options}
+        />
+      ))}
+    </Tab.Navigator>
   );
 };
