+++ src/types/components/providers.ts
@@ -1,3 +1,9 @@
 import { PropsWithChildren } from 'react';
 
-export type NavigationProviderProps = PropsWithChildren<{}>;
+export type NavigationProviderProps = PropsWithChildren<{
+
+}>;
+
+export type HttpProviderProps = PropsWithChildren<{
+    
+}>;
